improve notification api
