+++ src/main/java/net/minestom/server/adventure/audience/PacketGroupingAudience.java
@@ -10,6 +10,7 @@ import net.kyori.adventure.sound.SoundStop;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.title.TitlePart;
 import net.minestom.server.MinecraftServer;
+import net.minestom.server.advancements.Notification;
 import net.minestom.server.adventure.AdventurePacketConvertor;
 import net.minestom.server.coordinate.Point;
 import net.minestom.server.entity.Player;
@@ -28,7 +29,6 @@ import java.util.Collection;
  * An audience implementation that sends grouped packets if possible.
  */
 public interface PacketGroupingAudience extends ForwardingAudience {
-
     /**
      * Creates a packet grouping audience that copies an iterable of players. The
      * underlying collection is not copied, so changes to the collection will be
@@ -127,6 +127,15 @@ public interface PacketGroupingAudience extends ForwardingAudience {
     @Override
     default void stopSound(@NotNull SoundStop stop) {
         sendGroupedPacket(AdventurePacketConvertor.createSoundStopPacket(stop));
+    }
+
+    /**
+     * Send a {@link Notification} to the audience.
+     * @param notification the {@link Notification} to send
+     */
+    default void sendNotification(@NotNull Notification notification) {
+        sendGroupedPacket(notification.buildAddPacket());
+        sendGroupedPacket(notification.buildRemovePacket());
     }
 
     @Override
